solana: transfer: remove race condition
In the previous implementation, the outbox item account was a PDA keyed
by the sequence number. This meant that the client had to query the
next sequence, then submit a transaction with the outbox item account
seeded by that sequence. This created a race condition, since the query
and submit cannot be done atomically.

Instead, we just remove the requirement that the account is a PDA, and
instead require it to be a signer when creating the outbox item.
This removes the race condition because the client can just generate a
new unique address, and store it for later (which they had to do with
the sequence number anyway)