solana: Add check for no registered transceivers
Add guard on transfer instructions to prevent a transfer from being
processed when no transceivers are registered. This is done by checking
the `next_transceiver_id` field in the Config. This field
auto-increments when a transceiver is added. If this value is 0, then
the program is in a state where no transceivers have been registered.