solana: Add build, test, and lint targets to top-level dir
Adds Makefile targets for common Solana tasks in the top-level
directory. This allows the test and build pipeline to be run from the
main directory for both the EVM and Solana implementations.

This also helps to capture the Solana commands that are configured in
the .github/workflows/solana.yml file. In case this codebase is merged
or adapted into the Wormhole monorepo, the GitHub workflow may be
removed for modified. Therefore it is beneficial to track the
expected test configurations here.