wip: transfer before burn/mint to trigger hooks
wip because with this the test tx size is 5 bytes larger (for transfer)
than the limit. The way to mitigate is to introduce lookup tables.
That mitigation will exist in the tests, but will also have to be done
on real deployments (since the transfer hooks might add an arbitrary
number of extra acconuts, so we should leave as much headroom as possible)