wip: governance tests
these tests complain because the executable is being modified in the
transfer ownership instruction. this issue does not happen in devnet, or
solana-program-test, so maybe it's to do with the way anchor starts the
validator.

in any case, commenting out the upgrade authority transfer instruction,
the rest works. TODO: figure out a way to get around this