Makefile: Add Solana Makefile targets to top-level dir (#502)
* solana: Add build, test, and lint targets to top-level dir

Adds Makefile targets for common Solana tasks in the top-level
directory. This allows the test and build pipeline to be run from the
main directory for both the EVM and Solana implementations.

This also helps to capture the Solana commands that are configured in
the .github/workflows/solana.yml file. In case this codebase is merged
or adapted into the Wormhole monorepo, the GitHub workflow may be
removed for modified. Therefore it is beneficial to track the
expected test configurations here.

* solana: Update Makefile targets and clippy command

* Makefile: Remove all target

* Makefile: Update .PHONY target names

Co-authored-by: Andreas <41449730+nonergodic@users.noreply.github.com>

* Makefile: Add Solana prod build and general targets to build/test/lint

* CI: Update Forge fmt to use new Makefile target

* chore: Readd newlines

* solana: Add build, test, and lint targets to top-level dir

Adds Makefile targets for common Solana tasks in the top-level
directory. This allows the test and build pipeline to be run from the
main directory for both the EVM and Solana implementations.

This also helps to capture the Solana commands that are configured in
the .github/workflows/solana.yml file. In case this codebase is merged
or adapted into the Wormhole monorepo, the GitHub workflow may be
removed for modified. Therefore it is beneficial to track the
expected test configurations here.

* solana: Update Makefile targets and clippy command

* Makefile: Remove all target

* Makefile: Update .PHONY target names

Co-authored-by: Andreas <41449730+nonergodic@users.noreply.github.com>

* Makefile: Add Solana prod build and general targets to build/test/lint

* CI: Update Forge fmt to use new Makefile target

* chore: Readd newlines

* solana: Update Makefile to include CI commands

* evm: Add Makefile with build, tests, and lint targets

* CI: Update evm.yml to use new targets

* Makefile: Add generic build, test, and lint targets

* CI: Set evm.yml working directory to ./evm

* CI: Untar echidna to echidna-bin to avoid name conflicts

* CI: Fix untar location from PATH to echidna-bin

---------

Co-authored-by: John Saigle <john@asymmetric.re>
Co-authored-by: Andreas <41449730+nonergodic@users.noreply.github.com>