solana sdk: Fix getIsApproved to check inbox item
getIsApproved() was returning true when the VAA was posted, causing
track() to set the receipt state to DestinationInitiated prematurely.
It should check if the inbox item is initialized instead.

Fixes an issue that caused calls to NttManualRoute.complete() to
return early and never actually complete the transfer in rare cases
(receipt.DestinationInitiated === true and inbox item not created).